Workaround for odd SDL2 makefile bug breaking build pipeline
The generated makefile output complains about a missing make
target when run the first time with "make install" and fails.
Running it again "fixes" the issue somehow.

Not ideal, but the work around will not break if this suddendly
starts to work again for whatever reason. Make will just run
a second time and realize everything's done already.